Beautify experiment setup routine
The logic that determines whether the stuff inside .setup() runs once
or each time the experiment is instantiated (i.e., any time a request
hits the server) is now moved internal to the method. This cleans up
the experiment-specific file. The convention will be to have a .setup()
method that handles everything.