Fix a bug with Heroku
This is a temporary fix for an issue we were seeing with Heroku. When
you install Wallace using setuptools, it imports the version number
from inside the Wallace codebase. That triggers Wallace’s init file,
which imports the database, whose initialization depends on
SQL-Alchemy. That causes an error because sqlalchemy can’t be found.
For now, let’s avoid the morass.