Fix bug in waiting for Redis (#329)
To check if the Redis queue is available, we try to write to it.
However, sometimes the URL for the Redis queue is not immediately
available. The existing implementation retrieves this URL once and then
pings it over and over. The new implementation retrieves the URL once
per ping. Thus it is updated as soon as the URL becomes available,
preventing an infinite hang.