Address some of @davisagli's comments on config work
This allows config data loading to be marked as strict, which raises an error for unknown keys. This is intended to be used for the configuration files, but is currently disabled due to ongoing psiturk dependency.

It also adds dict access to the configuration object and enables loading of variables from the environment.