Load configuration even without local config.txt
Previously, when a local config.txt cannot be found in the current
working directory, a value error is raised. This change allows the
configuration to load even when there is no such config.txt, which
makes it possible to use to configuration even outside the context of
an experiment directory (e.g., to get AWS keys to load data from S3 in
an analysis script).