Be quieter about invalid configuration keys
Users will routinely have environment variables set that are not valid
keys for Dallinger's configuration. Rather than spew a list of these
each time an experiment is run, default to being quiet about this
unless the logger level is set to debug. On the other hand, if the user
includes an invalid configuration variable in a config file, then it
makes sense to complain.