Clean config [WIP] (#447)
* Remove unused task parameters

* Remove launch_in_sandbox_mode

Use `mode` instead.

* Remove login_username and login_pw

* Add missing mode

* Shorten "Database Parameters" heading

* Shorten "HIT Configuration" heading

* Shorten "Experiment Configuration" heading

* Use new config in Snake demo

* Update 2048 demo to use new config

* Shorten "Server Parameters" heading

* Rename settings => parameters

* Write some documentation on configuration

* Make config variables accessible as attributes

* Be quieter about invalid configuration keys

Users will routinely have environment variables set that are not valid
keys for Dallinger's configuration. Rather than spew a list of these
each time an experiment is run, default to being quiet about this
unless the logger level is set to debug. On the other hand, if the user
includes an invalid configuration variable in a config file, then it
makes sense to complain.

* Remove unused configuration variables

* Make 'team' a synonym of 'heroku_team'

(To improve backwards-compatibility with versions 2.x.)