Support Selenium drivers other than local PhantomJS
This change allows users to specify the webdriver_url configuration parameter, which points at a Selenium Server, such as http://192.168.1.2:4444/wd/hub. If this is specified the webdriver.Remote backend is used.
In addition, webdriver_type determines the browser being used. If a URL is specified these are passed as desired_capabilities to the hub server, otherwise it is used to pick a local webdriver. Currently supported values are 'phantomjs' and 'firefox'.