Wait after recruitment is complete
In some experiments, the experiment becomes complete the moment the
final participant submits their HIT, setting their status away from
`working`. Sometimes there is so little time between when the
experiment becomes complete and the server is shut off that the server
becomes unavailable *during* the interchange when the final bot
participant submits the questionnaire, leaving them unable to finish.
This change adds the delay before ever checking whether the experiment
is complete.