Fix sort order issue in MCMCP demo
When running a long chain, @jcpeterson noticed that every so often the
mapping between the "left" and "right" buttons flipped. The error was
non-deterministic and hard to predict. After a long debugging session,
we realized that the order of infos returned by the
`node/<node_id>/infos` route is non-deterministic (or at least not
always sequential in node id). This is a fix that behaves properly no
matter what order the returned infos are sorted in.

Closes #570.