Patch jupyter output stream
The ipykernel's implementation of iostream marks files as not writable. This makes
some sense, as they do not implement `fileno\(\)`, so can't be written by `os.write\(\)`
but it can cause libraries like pexpect to refuse to write data directly, which
still works. This patch overrides the value, and we will handle the special-casing
in different commits.