Show error if heroku login can't be completed rather than failing
If running through Jupyter, the heroku login can't be completed as
it requires using `pexpect.Process().interact()` but the stdout
we have isn't a real pty. This changes the logic of this function
to detect if the user is logged in, and if not to return appropriate
error messages if the user can't be prompted directly.