Wrap subprocess calls to redirect to nonfunctional std streams
Under Jupyter the stderr and stdout streams do not have file numbers,
so the subprocess family of calls cannot write to them directly. This
wraps these calls with a method to detect when calls do not specify
redirection of these streams and if it detects a nonfunctional output
it uses a temporary file and reads the data afterwards.

This causes output under Jupyter to be slower at updating than on
the command line, but allows for experiments to be run.