Make ad and consent pages use template inheritence
This adds a new default/ad.html and default/consent.html that represent the default versions of those pages
with block hooks to allow inheritence and overrides of specifics. Of course, full versions can still be provided.

This also removes on-the-fly modification of the ad template to insert missing variables, and instead
updats the templates themselves.