Improvements for repeatable test coverage
The `notify()` method of command_line.py is partially covered depending
on dictionary ordering, which causes spurious -0.19% test coverage deltas
in some builds. This adds a test for another use of that function that
although still being ordering dependent will fix the partial coverage
of that function in all cases.