Attempt to open experiment with clean profiles
The webbrowser module has implementations for opening windows in
common browsers, with the correct command line flags to open new
windows or new tabs. This checks if Chrome or Firefox is installed
and if so instantiates a new browser control that opens a new profile
using a temporary directory. If neither are available it falls back
to the ordering in webbrowser, so users who don't have Chrome or FF
see the existing behaviour.