Tag builds not on master with their git commit
This is intended to prevent the automatic releases of every single
master commit to the test PyPI server from preempting the real
release of a tagged version. When running setup() the script will
attempt to find the commit hash of the current HEAD and of
origin/master. If they don't match the build will be tagged as dev
with the commit hash appended.