Update quorum channel when generating status report
When running an experiment the controlling process periodically
requests a status report. This suborns that mechanism to also
publish periodic updates on the state of the quorum. This prevents
bots who lose their connection to the quorum channel over the critical
point of quorum being reached from being stuck in the waiting room.