Special case MacOS when opening new Chrome browsers
MacOS uses applescript to open its browsers as the Chrome (and other)
executables aren't in PATH. If the end user didn't set PATH to include
the MacOS subdirectory of the application package then opening a new
chrome would fall back to the webbrowser applescript method which
would not cause a new profile to be created.

This checks for the presence of the Chrome package (with a config
override for location) and uses that path on darwin in it is found.