Add finalize_hit to recruiter API
This allows recruiters to behave slightly statefully when assigning
payment to end users. Prior to this, the recruiter would pay the base
fee, then pay the bonus if there was one. Some recruitment services
only allow for a single payment to be made to a user, so the finalize
call enables a single payment to be made after both base payment and
bonus calls have been made.