Tox should use development mode installations
Note, the following contains supposition:

Tox uses sdists to install the package under test, but now we have a
pytest plugin our code is loaded earlier than test discovery, which
is too early for the path it's installed under to be whitelisted
by the coverage plugin.

This change instructs tox to use a development install instead, so
a direct link with the files on disk is used.