Move docker dependency out of extra
Having that dependency as an extra was causing problems when mounting the current
dallinger directory from the host inside the container.

Scenario: I'm developing Dallinger, and I want to use `dallinger docker debug`.
The local Dallinger directory is mounted as a volume into the docker container,
so that the container runs the code that is currently on the filesystem.

Before this change that caused an error, since the file dallinger.egg-info/requires.txt
from the host would include `docker` and `docker-compose`, but the dallinger image was lacking those.

The fix is this commit is just meant as a temporary patch.

A proper fix would be to separate Dallinger CLI from its server and worker and have two distinct python
packages for those.