Do not capture stdout/stderr during docker test runs
It's very useful to be able to see the timestamp github actions
provide for each line in the log, to reason about how the timeouts
impacted on the test.
pytest by default swallows all output during test runs.
Adding `-s` shows the output as they are run, instead of in a single
block after they run.