Pin Flask<2.0
We currently depend on sphinx-js for our documentation, which in turn specifies
Jinja2>2.0,<3.0
as a dependency. The new Flask 2.0 (released in May 2021) depends on Jinja>=3.

Ideally the dependencies we use to generate documentation should not impact the ones
used to run a web server, but for now they do.
Also upgrading to Flask 2.0 might require significant effort.