Use the current dallinger version's requirements.txt to constaint te `generate-constraints` command
The `generate-constraints` command should make sure the version it produces (especially those for
packages that require compilation, like psycopg) match the ones bundled in the docker image for
the dallinger version the package is pinned for.

This commit makes dallinger use the current (CLI) version for itself, and the requirements.txt file
that was produced for that version and used to build the corresponding docker image.