Minor update to network visualistion (#3932)
I have fixed a couple of small formatting issues with the dashboard network visualisation.

First, I have updated the text by Infos so that it says the class name rather than just 'info'. This is particularly useful for PsyNet, where the term 'Info' is not so informative.

Second, I have fixed an issue for the display of networks that have no nodes; previously the visualization would draw a line between these networks and an unrelated node from a different network. Now the network is just left without any connections.